// src/interfaces/App/Productos.interfaces.ts
export interface CuentaCorresponsalDetalle {
  nombre: string;
  cuenta: string;
  direccion: string;
  nroDocumento: string;
  moneda: string;
  tipoProducto: string;
  estado: string;
  sucursal: string;
  saldo: number;
  ctnro: string;
  clienteUid: string;
  telefonoCelular: string;
  correoElectronico: string;
}